*/mpconfigport.h: Use stdint.h to simplify things.
Avoids reinventing the wheel.  Or in this case, avoids reinventing the
pointer-sized-integer intptr_t and friends.

+++ ports/qemu-arm/mpconfigport.h
@@ -42,8 +42,8 @@
 #define UINT_FMT "%lu"
 #define INT_FMT "%ld"
 
-typedef int32_t mp_int_t; // must be pointer size
-typedef uint32_t mp_uint_t; // must be pointer size
+typedef intptr_t mp_int_t; // must be pointer size
+typedef uintptr_t mp_uint_t; // must be pointer size
 typedef long mp_off_t;
 
 // We need to provide a declaration/definition of alloca()
